Fix drag/drop shim z-index
When the CalendarPanel is floated, the shimming shows above message boxes. Window z-index management in Ext 3.x is hard to work around so the shim container was moved from body to the actual container of the calendar. That way the z-index for shimming is relative only to the container, not to everything in the document.