Fix for null error on calendar destroy
There is a very specific bug in Ext 4.1-b1 that only happens if you specify a renderSelector and a childEl with the same name. The Extensible StatusProxy class extends Ext.dd.StatusProxy and added a renderSelector that duplicated a childEl defined by it. It doesn't affect anything, but did expose this bug in Component.destroy(), which should check for null els anyway before operating on them.

I've submitted a patch to Sencha, but this code change will work around it with no functional difference to the calendar.