Fix bug after "add" action returns exception
The existing code only attempted to save if the added record did not already exist in the store. However, if the initial add attempt fails and is retried, the record will still be in the store (and doesn't need to be re-added) but the store save should still be executed. Rearranged the logic to attempt save every time.