Implementation for the bug: [OPEN-334] Month view: "+X more..." link missing. Details about the bug can be reviewed at: http://ext.ensible.com/forum/viewtopic.php?f=3&t=613&sid=9f1e63e3221d5e8bb110e99d2f4f3f5e
Implementation is to recalculate the no of events to display when the month is changed.

Steps to verify:
in the current TestApp example, load the current month view for march 2013 it has all the test events.  Click the < button to go to feb, then resize to make the browser smaller, then click > to go back to march, the + more links should be displayed properly (not hidden).