Fix incorrect day view start date shift
There are a bunch of spots in the view rendering code that get a 12 hour shift applied to avoid daylight savings edge cases. However, the underlying start date of the calendar should not have been shifted. In most cases the view logic was defaulting this start date back to 0, which was hiding the bug, but the day view was simply defaulting to the (incorrect) calendar start date. Removed the 12 hour shift from the start date that should never have been applied in the first place.